Very structured looking generation works

+++ Assets/Scripts/Island/Noise.cs
@@ -4,13 +4,14 @@ using UnityEngine;
 
 public static class Noise  {
 
-    public static float [,] GenerateNoiseMap(int mapWidth, int mapHeight, int seed, float scale, int octaves, float persistance, float lacunarity, Vector2 offset)
+    public static float [,] GenerateNoiseMap(float mapWidth, float mapHeight, int seed, float scale, int octaves, float persistance, float lacunarity, Vector2 offset)
     {
         //create noisemap array
-        float[,] noiseMap = new float[mapWidth, mapHeight];
+        float[,] noiseMap = new float[(int)mapWidth, (int)mapHeight];
 
         //pseudo random number generator
         System.Random prng = new System.Random(seed);
+
         Vector2[] octaveOffsets = new Vector2[octaves];
         for(int i = 0; i < octaves; i++)
         {
